Added LEDBar and servomotor tests

+++ projet/Library/lib_dir/LEDBar.cpp
@@ -29,7 +29,7 @@ void LEDBar::displayState(){
     *portPtr_ |= (1 << MR_); /* Memory mode on */
     *portPtr_ &= ~(1 << OE_); /* Display mode on */
     for(uint8_t i = MAX_BIT_POSITION; i > MIN_BIT_POSITION; i--){
-        if(LEDState_ & (1 << (i - 1))){
+        if(LEDState_ & ( (uint32_t)1 << (i - 1))){
             *portPtr_ |= (1 << DS_); /* DS */
         }
         else{
@@ -39,7 +39,6 @@ void LEDBar::displayState(){
         *portPtr_ &= ~(1 << SHC_);
         *portPtr_ |= (1 << SHC_);
         *portPtr_ &= ~(1 << SHC_);
-        //1
     }
     
     *portPtr_ &= ~(1 << STC_);
@@ -54,7 +53,7 @@ void LEDBar::openDoor(uint8_t doorType){
     Debug debug = Debug();
     for(uint8_t i = MAX_LED_BAR_POSITION; i > MIN_BIT_POSITION; i--){
         
-        LEDState_ &= ~(1 << ((MAX_LED_BAR_POSITION * doorType) + (i - 1)));
+        LEDState_ &= ~((uint32_t)1 << ((MAX_LED_BAR_POSITION * doorType) + (i - 1)));
         displayState();
         _delay_ms(500);
     }
@@ -64,7 +63,7 @@ void LEDBar::closeDoor(uint8_t doorType)
 {
     for(uint8_t i = MIN_BIT_POSITION; i < MAX_LED_BAR_POSITION; i++)
     {
-        LEDState_ |= (1 << (MAX_LED_BAR_POSITION * doorType + i));
+        LEDState_ |= ((uint32_t)1 << (MAX_LED_BAR_POSITION * doorType + i));
         displayState();
         _delay_ms(500);
     }
